the whole as mod

+++ src/main/java/dev/sterner/better_stuck_arrows/StuckSpectralArrowsFeatureRenderer.java
@@ -12,7 +12,6 @@ import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.projectile.ArrowEntity;
 import net.minecraft.entity.projectile.SpectralArrowEntity;
 import net.minecraft.util.math.MathHelper;
 
@@ -35,8 +34,8 @@ public class StuckSpectralArrowsFeatureRenderer<T extends LivingEntity, M extend
     protected void renderObject(MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, Entity entity, float directionX, float directionY, float directionZ, float tickDelta) {
         float f = MathHelper.sqrt(directionX * directionX + directionZ * directionZ);
         SpectralArrowEntity arrowEntity = new SpectralArrowEntity(entity.world, entity.getX(), entity.getY(), entity.getZ());
-        arrowEntity.setYaw((float) (Math.atan2((double) directionX, (double) directionZ) * 57.2957763671875));
-        arrowEntity.setPitch((float) (Math.atan2((double) directionY, (double) f) * 57.2957763671875));
+        arrowEntity.setYaw((float) (Math.atan2(directionX, directionZ) * 57.2957763671875));
+        arrowEntity.setPitch((float) (Math.atan2(directionY, f) * 57.2957763671875));
         arrowEntity.prevYaw = arrowEntity.getYaw();
         arrowEntity.prevPitch = arrowEntity.getPitch();
         this.dispatcher.render(arrowEntity, 0.0, 0.0, 0.0, 0.0F, tickDelta, matrices, vertexConsumers, light);
